Fixing the examples
